allow a maxiter parameter to be float

+++ scripts/exp_recsys_sp.py
@@ -520,7 +520,7 @@ def command_line_parser():
     ap.add_argument('-k', '--dim', dest='k', type=int, default=1)
     ap.add_argument('--alpha', dest='alpha', type=float, default=1.0)
     ap.add_argument('--tol', type=float, default=1e-05)
-    ap.add_argument('--maxiter', type=int, default=200)
+    ap.add_argument('--maxiter', type=float, default=200)
 
     # parsing
     opt = ap.parse_args()
